+++ src/web/web_socket.ts
@@ -1,11 +1,12 @@
 import { CpCall, trans, RpcFrame, RpcFrameCtrl } from "cpcall";
-import { PassiveDataCollector } from "evlib/async";
+import { DataCollector } from "evlib/async";
 
 function webSocketToIter(webSocket: WebSocket) {
-  const collector = new PassiveDataCollector<RpcFrame, Error | void>();
+  const collector = new DataCollector<RpcFrame, Error | void>();
   webSocket.addEventListener("message", (e) => {
     if (e.data instanceof ArrayBuffer) {
-      collector.yield(trans.decodeCpcFrame(new Uint8Array(e.data)).frame);
+      const framesIterator = trans.unpackCpcFrames(new Uint8Array(e.data), 0);
+      for (const frame of framesIterator) collector.yield(frame);
     }
   });
   webSocket.addEventListener("close", (e) => {
@@ -14,7 +15,7 @@ function webSocketToIter(webSocket: WebSocket) {
   webSocket.addEventListener("error", (e) => {
     collector.close(new Error("Websocket encountered an error", { cause: (e as any).data }));
   });
-  return collector.getAsyncGen();
+  return collector;
 }
 
 /** 创建一个基于 WebSocket 的 CpCall 实例。WebSocket 的状态必须是 open。 否则抛出异常
@@ -30,12 +31,20 @@ class WsRpcFrameCtrl implements RpcFrameCtrl {
     this.frameIter = webSocketToIter(ws);
   }
   frameIter: AsyncIterable<RpcFrame>;
-  sendFrame(frame: RpcFrame): void {
-    //todo: 需要改进，当源关闭后直接将 callee 和 caller 只为
-    if (this.ws.readyState == this.ws.OPEN) this.ws.send(new trans.CpcFrameEncoder(frame).encode());
-    else {
+  private link: RpcFrame[] = [];
+  send() {
+    //todo: 需要改进，当源关闭后直接将 callee 和 caller 关闭
+    if (this.ws.readyState === this.ws.OPEN) {
+      const chunk = trans.packCpcFrames(this.link);
+      this.ws.send(chunk);
+      this.link.length = 0;
+    } else {
       this.dispose();
     }
+  }
+  sendFrame(frame: RpcFrame): void {
+    if (this.link.length === 0) Promise.resolve().then(() => this.send());
+    this.link.push(frame);
   }
   close() {
     this.ws.close();
@@ -52,8 +61,8 @@ interface WebSocket {
   close(): void;
   send(data: Uint8Array): void;
   addEventListener(name: "message", fn: (e: { readonly data: any }) => void): void;
-  addEventListener(name: "close", fn: (e: SameEvent) => void): void;
-  addEventListener(name: "error", fn: (e: SameEvent) => void): void;
+  addEventListener(name: string, fn: (e: SameEvent) => void): void;
+  addEventListener(name: string, fn: (e: SameEvent) => void): void;
 }
 interface SameEvent {
   [key: string]: any;