+++ test/cpcall/ws-cpc.test.ts
@@ -1,0 +1,65 @@
+import { createWebSocketCpc } from "cpcall/web";
+import { vi, test, expect } from "vitest";
+type WS = Parameters<typeof createWebSocketCpc>[0];
+
+class MockWebSocket extends EventTarget implements WS {
+  constructor() {
+    super();
+  }
+  readonly OPEN = 1;
+  binaryType = "ArrayBuffer";
+  send: (data: Uint8Array) => void = vi.fn();
+  close: () => void = vi.fn();
+  readyState = this.OPEN;
+}
+interface MockWebSocket extends EventTarget {
+  addEventListener(name: "message", fn: (e: { readonly data: any }) => void): void;
+  addEventListener(name: string, fn: (e: Event) => void): void;
+}
+function createConnectedWs() {
+  const ws1 = new MockWebSocket();
+  const ws2 = new MockWebSocket();
+  function send(ws: MockWebSocket, frame: any) {
+    const event = new Event("message") as Event & { data: ArrayBuffer | string };
+    let buf: ArrayBuffer | string;
+    if (frame instanceof Uint8Array) {
+      if (frame.buffer.byteLength === frame.byteLength) buf = frame.buffer;
+      else {
+        const u8Arr = new Uint8Array(frame);
+        buf = u8Arr.buffer;
+      }
+    } else buf = String(frame);
+    event.data = buf;
+
+    ws.dispatchEvent(event);
+  }
+
+  ws1.send = (frame) => send(ws2, frame);
+  ws1.close = function () {
+    this.dispatchEvent(new Event("close"));
+    ws2.dispatchEvent(new Event("close"));
+  };
+
+  ws2.send = (frame) => send(ws1, frame);
+  ws2.close = function () {
+    this.dispatchEvent(new Event("close"));
+    ws1.dispatchEvent(new Event("close"));
+  };
+  return { ws1, ws2 };
+}
+test("事件触发", async function () {
+  const { ws1, ws2 } = createConnectedWs();
+  const cpc1 = createWebSocketCpc(ws1);
+  const cpc2 = createWebSocketCpc(ws2);
+
+  const fn = vi.fn((arg) => arg);
+  cpc2.setFn("abc", fn);
+  const cal = cpc1.caller;
+  const res = await Promise.all([cal.call("abc", 1), cal.call("abc", 3)]);
+  expect(res).toEqual([1, 3]);
+});
+test("错误的状态", function () {
+  const ws = new MockWebSocket();
+  ws.readyState = 0;
+  expect(() => createWebSocketCpc(ws), "websocket状态必须为已连接").toThrowError();
+});