+++ test/extra/remote_proxy.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi, beforeEach } from "vitest";
 import { RpcFrame, CpCall } from "cpcall";
-import { PassiveDataCollector } from "evlib/async";
+import { DataCollector } from "evlib/async";
 
 interface SubCall {
   ab: number;
@@ -77,7 +77,7 @@ test("exec", function () {
 
 function createCpc() {
   const onSendFrame = vi.fn();
-  let hd = new PassiveDataCollector<RpcFrame>();
-  let cpc = new CpCall({ frameIter: hd.getAsyncGen(), sendFrame: onSendFrame });
+  let hd = new DataCollector<RpcFrame>();
+  let cpc = new CpCall({ frameIter: hd, sendFrame: onSendFrame });
   return { onSendFrame, hd, cpc };
 }